requirements.txt
litellm
deepgram-sdk
playsound
langchain-groq 
langchain_google_genai
langchain_chroma
chromadb
Pillow
pyperclip
pyaudio
python-dotenv
instructor 
pydantic
colorama
tavily-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client